chore: Unified bbox checking for selection
