+++ packages/element/src/selection.ts
@@ -1,10 +1,4 @@
 import { arrayToMap, isShallowEqual, type Bounds } from "@excalidraw/common";
-import {
-  lineSegment,
-  pointFrom,
-  pointRotateRads,
-  type GlobalPoint,
-} from "@excalidraw/math";
 
 import type {
   AppState,
@@ -15,18 +9,14 @@ import type {
 import {
   boundsContainBounds,
   doBoundsIntersect,
-  elementCenterPoint,
+  elementsOverlappingBBox,
   getElementAbsoluteCoords,
   getElementBounds,
-  pointInsideBounds,
 } from "./bounds";
-import { intersectElementWithLineSegment } from "./collision";
 import { isElementInViewport } from "./sizeHelpers";
 import {
-  isArrowElement,
   isBoundToContainer,
   isFrameLikeElement,
-  isFreeDrawElement,
   isLinearElement,
   isTextElement,
 } from "./typeChecks";
@@ -38,7 +28,6 @@ import {
 
 import { LinearElementEditor } from "./linearElementEditor";
 import { selectGroupsForSelectedElements } from "./groups";
-import { getBoundTextElement } from "./textElement";
 
 import type {
   ElementsMap,
@@ -87,6 +76,37 @@ export const excludeElementsInFramesFromSelection = <
   return excludeElementsFromFrames(selectedElements, framesInSelection);
 };
 
+// Correct for frame clipping: elementsOverlappingBBox uses unclipped element
+// bounds, but selection should only consider the visible (frame-clipped)
+// portion of elements in frames.
+const getFrameClippedAABB = (
+  element: NonDeletedExcalidrawElement,
+  elementsMap: ElementsMap,
+): Bounds | null => {
+  const associatedFrame = getContainingFrame(element, elementsMap);
+  if (
+    !associatedFrame ||
+    !elementOverlapsWithFrame(element, associatedFrame, elementsMap)
+  ) {
+    return null;
+  }
+  const strokeWidth = element.strokeWidth;
+  let elementAABB = getElementBounds(element, elementsMap);
+  elementAABB = [
+    elementAABB[0] - strokeWidth / 2,
+    elementAABB[1] - strokeWidth / 2,
+    elementAABB[2] + strokeWidth / 2,
+    elementAABB[3] + strokeWidth / 2,
+  ] as Bounds;
+  const frameAABB = getElementBounds(associatedFrame, elementsMap);
+  return [
+    Math.max(elementAABB[0], frameAABB[0]),
+    Math.max(elementAABB[1], frameAABB[1]),
+    Math.min(elementAABB[2], frameAABB[2]),
+    Math.min(elementAABB[3], frameAABB[3]),
+  ] as Bounds;
+};
+
 export const getElementsWithinSelection = (
   elements: readonly NonDeletedExcalidrawElement[],
   selection: NonDeletedExcalidrawElement,
@@ -107,31 +127,11 @@ export const getElementsWithinSelection = (
     selectionX2,
     selectionY2,
   ] as Bounds;
-  const selectionEdges = [
-    lineSegment<GlobalPoint>(
-      pointFrom(selectionX1, selectionY1),
-      pointFrom(selectionX2, selectionY1),
-    ),
-    lineSegment<GlobalPoint>(
-      pointFrom(selectionX2, selectionY1),
-      pointFrom(selectionX2, selectionY2),
-    ),
-    lineSegment<GlobalPoint>(
-      pointFrom(selectionX2, selectionY2),
-      pointFrom(selectionX1, selectionY2),
-    ),
-    lineSegment<GlobalPoint>(
-      pointFrom(selectionX1, selectionY2),
-      pointFrom(selectionX1, selectionY1),
-    ),
-  ];
 
-  const framesInSelection = excludeElementsInFrames
-    ? new Set<NonDeletedExcalidrawElement["id"]>()
-    : null;
   const groups: Record<string, NonDeletedExcalidrawElement[]> = {};
-  const elementsInSelection: Set<NonDeletedExcalidrawElement> = new Set();
 
+  // First pass: filter selectable elements and build groups map
+  const selectableElements: NonDeletedExcalidrawElement[] = [];
   for (const element of elements) {
     if (shouldIgnoreElementFromSelection(element)) {
       continue;
@@ -147,188 +147,62 @@ export const getElementsWithinSelection = (
       groups[groupId].push(element);
     }
 
-    const strokeWidth = element.strokeWidth;
-    let labelAABB: Bounds | null = null;
-    let elementAABB = getElementBounds(element, elementsMap);
-
-    elementAABB = [
-      elementAABB[0] - strokeWidth / 2,
-      elementAABB[1] - strokeWidth / 2,
-      elementAABB[2] + strokeWidth / 2,
-      elementAABB[3] + strokeWidth / 2,
-    ] as Bounds;
-
-    // Whether the element bounds should include the bound text element bounds
-    const boundTextElement =
-      isArrowElement(element) && getBoundTextElement(element, elementsMap);
-    if (boundTextElement) {
-      const { x, y } = LinearElementEditor.getBoundTextElementPosition(
-        element,
-        boundTextElement,
-        elementsMap,
-      );
-      labelAABB = [
-        x,
-        y,
-        x + boundTextElement.width,
-        y + boundTextElement.height,
-      ] as Bounds;
-    }
-
-    // Clip element bounds by its containing frame (if any), since only the
-    // visible (frame-clipped) portion of the element is relevant for selection.
-    const associatedFrame = getContainingFrame(element, elementsMap);
-    if (
-      associatedFrame &&
-      elementOverlapsWithFrame(element, associatedFrame, elementsMap)
-    ) {
-      const frameAABB = getElementBounds(associatedFrame, elementsMap);
-      elementAABB = [
-        Math.max(elementAABB[0], frameAABB[0]),
-        Math.max(elementAABB[1], frameAABB[1]),
-        Math.min(elementAABB[2], frameAABB[2]),
-        Math.min(elementAABB[3], frameAABB[3]),
-      ] as Bounds;
-
-      labelAABB = labelAABB
-        ? ([
-            Math.max(labelAABB[0], frameAABB[0]),
-            Math.max(labelAABB[1], frameAABB[1]),
-            Math.min(labelAABB[2], frameAABB[2]),
-            Math.min(labelAABB[3], frameAABB[3]),
-          ] as Bounds)
-        : null;
-    }
-
-    const commonAABB = labelAABB
-      ? ([
-          Math.min(labelAABB[0], elementAABB[0]),
-          Math.min(labelAABB[1], elementAABB[1]),
-          Math.max(labelAABB[2], elementAABB[2]),
-          Math.max(labelAABB[3], elementAABB[3]),
-        ] as Bounds)
-      : elementAABB;
-
-    // ============== Evaluation ==============
-
-    // 1. If the selection box WRAPs the element's AABB, then add it to the
-    //    selection and move on, regardless of the selection mode.
-    //
-    //    PERF: This trick only works with axis-aligned box selection and the
-    //          current convex element shapes!
-    if (boundsContainBounds(selectionBounds, commonAABB)) {
-      if (framesInSelection && isFrameLikeElement(element)) {
-        framesInSelection.add(element.id);
-      }
-      elementsInSelection.add(element);
-      continue;
-    }
-
-    // 2. Handle the case where the label is overlapped by the selection box
-    if (
-      boxSelectionMode === "overlap" &&
-      labelAABB &&
-      doBoundsIntersect(selectionBounds, labelAABB)
-    ) {
-      elementsInSelection.add(element);
-      continue;
-    }
-
-    // 3. Handle the case where the selection is not wrapping the element, but
-    //    it does intersect the element's outline (non-AABB).
-    if (
-      boxSelectionMode === "overlap" &&
-      doBoundsIntersect(selectionBounds, elementAABB)
-    ) {
-      let hasIntersection = false;
-
-      // Preliminary check potential intersection imprecision
-      if (isLinearElement(element) || isFreeDrawElement(element)) {
-        const center = elementCenterPoint(element, elementsMap);
-        hasIntersection = element.points.some((point) => {
-          const rotatedPoint = pointRotateRads(
-            pointFrom<GlobalPoint>(element.x + point[0], element.y + point[1]),
-            center,
-            element.angle,
-          );
-
-          return pointInsideBounds(rotatedPoint, selectionBounds);
-        });
-      } else {
-        const nonRotatedElementBounds = getElementBounds(
-          element,
-          elementsMap,
-          true,
-        );
-        const center = elementCenterPoint(element, elementsMap);
-        hasIntersection = [
-          pointRotateRads(
-            pointFrom<GlobalPoint>(
-              (nonRotatedElementBounds[0] + nonRotatedElementBounds[2]) / 2,
-              nonRotatedElementBounds[1],
-            ),
-            center,
-            element.angle,
-          ),
-          pointRotateRads(
-            pointFrom<GlobalPoint>(
-              nonRotatedElementBounds[2],
-              (nonRotatedElementBounds[1] + nonRotatedElementBounds[3]) / 2,
-            ),
-            center,
-            element.angle,
-          ),
-          pointRotateRads(
-            pointFrom<GlobalPoint>(
-              (nonRotatedElementBounds[0] + nonRotatedElementBounds[2]) / 2,
-              nonRotatedElementBounds[3],
-            ),
-            center,
-            element.angle,
-          ),
-          pointRotateRads(
-            pointFrom<GlobalPoint>(
-              nonRotatedElementBounds[0],
-              (nonRotatedElementBounds[1] + nonRotatedElementBounds[3]) / 2,
-            ),
-            center,
-            element.angle,
-          ),
-        ].some((point) => {
-          return pointInsideBounds(
-            pointRotateRads(point, center, element.angle),
-            selectionBounds,
-          );
-        });
-      }
-
-      if (!hasIntersection) {
-        hasIntersection = selectionEdges.some(
-          (selectionEdge) =>
-            intersectElementWithLineSegment(
-              element,
-              elementsMap,
-              selectionEdge,
-              strokeWidth / 2,
-              true, // Stop at first hit for better performance
-            ).length > 0,
-        );
-      }
-
-      if (hasIntersection) {
-        if (framesInSelection && isFrameLikeElement(element)) {
-          framesInSelection.add(element.id);
-        }
-
-        elementsInSelection.add(element);
-        continue;
-      }
-    }
-
-    // 4. We don't need to handle when the selection is inside the element
-    //    as it is separately handled in App.
+    selectableElements.push(element);
   }
 
+  // Use elementsOverlappingBBox for bounds checking (without frame clipping)
+  let elementsInSelectionArr = elementsOverlappingBBox({
+    elements: selectableElements,
+    elementsMap,
+    bounds: selectionBounds,
+    type: boxSelectionMode,
+  });
+
+  if (boxSelectionMode === "overlap") {
+    // Remove elements whose frame-clipped bounds no longer overlap the selection
+    elementsInSelectionArr = elementsInSelectionArr.filter((element) => {
+      const clippedAABB = getFrameClippedAABB(element, elementsMap);
+      if (!clippedAABB) {
+        return true;
+      }
+      return (
+        boundsContainBounds(selectionBounds, clippedAABB) ||
+        doBoundsIntersect(selectionBounds, clippedAABB)
+      );
+    });
+  } else {
+    // "contain" mode: add elements whose clipped bounds fit within the
+    // selection, but whose unclipped bounds did not (so elementsOverlappingBBox
+    // missed them)
+    const inSelectionSet = new Set(elementsInSelectionArr);
+    for (const element of selectableElements) {
+      if (inSelectionSet.has(element)) {
+        continue;
+      }
+      const clippedAABB = getFrameClippedAABB(element, elementsMap);
+      if (clippedAABB && boundsContainBounds(selectionBounds, clippedAABB)) {
+        elementsInSelectionArr.push(element);
+        inSelectionSet.add(element);
+      }
+    }
+  }
+
+  // Track frames in selection and build working set
+  const framesInSelection = excludeElementsInFrames
+    ? new Set<NonDeletedExcalidrawElement["id"]>()
+    : null;
+
+  const elementsInSelection = new Set(elementsInSelectionArr);
+
+  if (framesInSelection) {
+    for (const element of elementsInSelection) {
+      if (isFrameLikeElement(element)) {
+        framesInSelection.add(element.id);
+      }
+    }
+  }
+
+  // Exclude frame children when their frame is also selected
   if (framesInSelection) {
     elementsInSelection.forEach((element) => {
       if (element.frameId && framesInSelection.has(element.frameId)) {
